Hoist NoneCalculation to its own file.
Step 3 of the Big Break Up, for exactly the same reasons as step 2.
