+++ PyRoute/Calculation/NoneCalculation.py
@@ -1,0 +1,35 @@
+"""
+Created on Aug 09, 2023
+
+@author: CyberiaResurrection
+"""
+from PyRoute.AllyGen import AllyGen
+from PyRoute.Calculation.RouteCalculation import RouteCalculation
+
+
+class NoneCalculation(RouteCalculation):
+    def __init__(self, galaxy):
+        super(NoneCalculation, self).__init__(galaxy)
+
+    # Pure HG Base jump distance cost
+    distance_weight = [0, 30, 50, 75, 130, 230, 490]
+
+    def generate_routes(self):
+        # self.generate_base_routes()
+        pass
+
+    def calculate_routes(self):
+        pass
+
+    def base_route_filter(self, star, neighbor):
+        if not AllyGen.are_owned_allies(star.alg_code, neighbor.alg_code):
+            return True
+        return False
+
+    def base_range_routes(self, star, neighbor):
+        pass
+
+    def route_weight(self, star, target):
+        dist = star.hex_distance(target)
+        weight = self.distance_weight[dist]
+        return weight
