+++ PyRoute/Galaxy.py
@@ -14,8 +14,9 @@ import math
 import networkx as nx
 
 from PyRoute.Star import Star
-from PyRoute.Calculation.TradeCalculation import TradeCalculation, NoneCalculation, CommCalculation, \
+from PyRoute.Calculation.TradeCalculation import TradeCalculation, CommCalculation, \
     OwnedWorldCalculation
+from PyRoute.Calculation.NoneCalculation import NoneCalculation
 from PyRoute.Calculation.XRouteCalculation import XRouteCalculation
 from PyRoute.StatCalculation import ObjectStatistics
 from PyRoute.AllyGen import AllyGen
